Add Dockerfile and docker-compose.yml
The dockerfile has two stages: build and app

The build stage, basically enables the project to run. It installs
hex, rebar and fetchs the dependencies. This step also can create a
release. For that, one needs to set the MIX_ENV=prod.

The app stage just copies the release from the previous step and run
it